Feat discount

+++ src/main/java/com/onulstore/service/QuestionService.java
@@ -1,9 +1,9 @@
 package com.onulstore.service;
 
 import com.onulstore.config.SecurityUtil;
-import com.onulstore.config.exception.Exception;
+import com.onulstore.config.exception.CustomException;
 import com.onulstore.domain.enums.Authority;
-import com.onulstore.domain.enums.ErrorResult;
+import com.onulstore.domain.enums.CustomErrorResult;
 import com.onulstore.domain.member.Member;
 import com.onulstore.domain.member.MemberRepository;
 import com.onulstore.domain.product.Product;
@@ -16,6 +16,7 @@ import com.onulstore.web.dto.QuestionDto;
 import java.time.LocalDateTime;
 import com.onulstore.web.dto.QuestionDto.QuestionResponse;
 import lombok.RequiredArgsConstructor;
+import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
@@ -35,10 +36,14 @@ public class QuestionService {
     // 질문 등록
     @Transactional
     public void insertQuestion(QuestionDto.QuestionRequest questionDto) {
-        Member member = memberRepository.findById(SecurityUtil.getCurrentMemberId()).orElseThrow(
-            () -> new Exception(ErrorResult.NOT_EXIST_USER));
+        if (SecurityContextHolder.getContext().getAuthentication().getPrincipal()
+            .equals("anonymousUser")) {
+            throw new CustomException(CustomErrorResult.LOGIN_NEEDED);
+        }
+        Member member = memberRepository.findById(SecurityUtil.getCurrentMemberId())
+            .orElseThrow(() -> new CustomException(CustomErrorResult.NOT_EXIST_USER));
         Product product = productRepository.findById(questionDto.getProductId()).orElseThrow(
-            () -> new Exception(ErrorResult.PRODUCT_NOT_FOUND));
+            () -> new CustomException(CustomErrorResult.PRODUCT_NOT_FOUND));
 
         Question question = Question.builder()
             .member(member)
@@ -54,16 +59,18 @@ public class QuestionService {
     // 질문 수정
     @Transactional
     public QuestionResponse updateQuestion(Long questionId, QuestionDto.QuestionRequest questionDto) {
-        Member member = memberRepository.findById(SecurityUtil.getCurrentMemberId()).orElseThrow(
-            () -> new Exception(ErrorResult.NOT_EXIST_USER));
+        if (SecurityContextHolder.getContext().getAuthentication().getPrincipal()
+            .equals("anonymousUser")) {
+            throw new CustomException(CustomErrorResult.LOGIN_NEEDED);
+        }
+        Member member = memberRepository.findById(SecurityUtil.getCurrentMemberId())
+            .orElseThrow(() -> new CustomException(CustomErrorResult.NOT_EXIST_USER));
         Product product = productRepository.findById(questionDto.getProductId()).orElseThrow(
-            () -> new Exception(ErrorResult.PRODUCT_NOT_FOUND));
-
+            () -> new CustomException(CustomErrorResult.PRODUCT_NOT_FOUND));
         Question question = questionRepository.findById(questionId).orElseThrow(
-            () -> new Exception(ErrorResult.NOT_EXIST_QUESTION));
-
+            () -> new CustomException(CustomErrorResult.NOT_EXIST_QUESTION));
         if (!member.getId().equals(question.getMember().getId())) {
-            throw new Exception(ErrorResult.USER_NOT_MATCH);
+            throw new CustomException(CustomErrorResult.USER_NOT_MATCH);
         }
         question.setTitle(questionDto.getTitle());
         question.setContent(questionDto.getContent());
@@ -74,14 +81,21 @@ public class QuestionService {
     // 질문 삭제
     @Transactional
     public void deleteQuestion(Long questionId) {
-        Member member = memberRepository.findById(SecurityUtil.getCurrentMemberId()).orElseThrow(
-            () -> new Exception(ErrorResult.NOT_EXIST_USER));
+        if (SecurityContextHolder.getContext().getAuthentication().getPrincipal()
+            .equals("anonymousUser")) {
+            throw new CustomException(CustomErrorResult.LOGIN_NEEDED);
+        }
+        Member member = memberRepository.findById(SecurityUtil.getCurrentMemberId())
+            .orElseThrow(() -> new CustomException(CustomErrorResult.NOT_EXIST_USER));
+        if (!member.getAuthority().equals(Authority.ROLE_ADMIN.getKey())) {
+            throw new CustomException(CustomErrorResult.ACCESS_PRIVILEGE);
+        }
 
         Question question = questionRepository.findById(questionId).orElseThrow(
-            () -> new Exception(ErrorResult.NOT_EXIST_QUESTION));
+            () -> new CustomException(CustomErrorResult.NOT_EXIST_QUESTION));
 
         if (!member.getId().equals(question.getMember().getId())) {
-            throw new Exception(ErrorResult.USER_NOT_MATCH);
+            throw new CustomException(CustomErrorResult.USER_NOT_MATCH);
         }
         questionRepository.deleteById(questionId);
     }
@@ -89,18 +103,22 @@ public class QuestionService {
     // 질문 상세 조회
     @Transactional
     public QuestionDto.QuestionRes getQuestion(Long productId, Long questionId) {
-        Member member = memberRepository.findById(SecurityUtil.getCurrentMemberId()).orElseThrow(
-            () -> new Exception(ErrorResult.NOT_EXIST_USER));
+        if (SecurityContextHolder.getContext().getAuthentication().getPrincipal()
+            .equals("anonymousUser")) {
+            throw new CustomException(CustomErrorResult.LOGIN_NEEDED);
+        }
+        Member member = memberRepository.findById(SecurityUtil.getCurrentMemberId())
+            .orElseThrow(() -> new CustomException(CustomErrorResult.NOT_EXIST_USER));
         Product product = productRepository.findById(productId).orElseThrow(
-            () -> new Exception(ErrorResult.PRODUCT_NOT_FOUND));
+            () -> new CustomException(CustomErrorResult.PRODUCT_NOT_FOUND));
         Question question = questionRepository.findById(questionId).orElseThrow(
-            () -> new Exception(ErrorResult.NOT_EXIST_QUESTION));
+            () -> new CustomException(CustomErrorResult.NOT_EXIST_QUESTION));
         QuestionAnswer questionAnswer = questionAnswerRepository.findByQuestionId(question.getId()).orElse(null);
 
         if (question.getSecret() == 'Y') {
             if (!(member.getId().equals(question.getMember().getId()) || member.getAuthority()
                 .equals(Authority.ROLE_ADMIN.getKey()))) {
-                throw new Exception(ErrorResult.SECRET_QUESTION);
+                throw new CustomException(CustomErrorResult.SECRET_QUESTION);
             }
             return QuestionDto.QuestionRes.of(question, questionAnswer);
         }
@@ -111,7 +129,7 @@ public class QuestionService {
     @Transactional
     public List<QuestionDto.QuestionResponse> getQuestionList(Long productId) {
         Product product = productRepository.findById(productId).orElseThrow(
-            () -> new Exception(ErrorResult.PRODUCT_NOT_FOUND));
+            () -> new CustomException(CustomErrorResult.PRODUCT_NOT_FOUND));
 
         List<Question> questions = questionRepository.findAllByProductId(product.getId());
         List<QuestionDto.QuestionResponse> questionList = new ArrayList<>();
@@ -125,8 +143,12 @@ public class QuestionService {
     // 질문 전체 조회(멤버별)
     @Transactional
     public List<QuestionDto.QuestionResponse> getMemberQuestionList() {
-        Member member = memberRepository.findById(SecurityUtil.getCurrentMemberId()).orElseThrow(
-            () -> new Exception(ErrorResult.NOT_EXIST_USER));
+        if (SecurityContextHolder.getContext().getAuthentication().getPrincipal()
+            .equals("anonymousUser")) {
+            throw new CustomException(CustomErrorResult.LOGIN_NEEDED);
+        }
+        Member member = memberRepository.findById(SecurityUtil.getCurrentMemberId())
+            .orElseThrow(() -> new CustomException(CustomErrorResult.NOT_EXIST_USER));
 
         List<Question> questions = questionRepository.findAllByMemberId(member.getId());
         List<QuestionDto.QuestionResponse> questionList = new ArrayList<>();
@@ -138,9 +160,19 @@ public class QuestionService {
     }
 
     @Transactional
-    public Integer questionDashBoard(LocalDateTime localDateTime) {
-        List<Question> questionList = questionRepository.findAllByCreatedDateAfter(localDateTime);
-        return questionList.size();
+    public Long questionDashBoard(LocalDateTime localDateTime) {
+        if (SecurityContextHolder.getContext().getAuthentication().getPrincipal()
+            .equals("anonymousUser")) {
+            throw new CustomException(CustomErrorResult.LOGIN_NEEDED);
+        }
+        Member member = memberRepository.findById(SecurityUtil.getCurrentMemberId())
+            .orElseThrow(() -> new CustomException(CustomErrorResult.NOT_EXIST_USER));
+        if (!member.getAuthority().equals(Authority.ROLE_ADMIN.getKey())) {
+            throw new CustomException(CustomErrorResult.ACCESS_PRIVILEGE);
+        }
+
+        Long questions = questionRepository.countByCreatedDateAfter(localDateTime);
+        return questions;
     }
 
 }
